Split out code and add dev command using nodemon (#9)
