+++ packages/name-api/package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "dev": "nodemon src/index.ts",
     "start": "ncc run src/index.ts",
     "build": "ncc build src/index.ts -sm -e fastify -e pino",
     "size": "size-limit",
@@ -38,6 +39,7 @@
   },
   "devDependencies": {
     "@vercel/ncc": "^0.36.0",
+    "nodemon": "^3.1.0",
     "ts-node": "^10.9.2"
   }
 }
